requirements.txt
asgiref==3.5.1
Django==4.0.4
django-environ==0.8.1
gunicorn==20.1.0
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.0.0
